refactor: 꾸미 요소 애니메이션 통합 hooks 추가(class 기반 구현 후 class 헬퍼 함수 경유)
