--- src/feature/PentagramDecoration/Boids/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './usePentagramBoidsAnimation'
